+++ RELEASE_NOTES.md
@@ -1,9 +1,14 @@
 # Change log
 
+## 0.12.10 (2021-05-07)
+
+* Escapes strings when generating/formatting code (unless they are in the command block)
+* Always unwraps `V_Optional` values during expression evaluation
+* Fixes issues with using expressions in placeholder option values
+
 ## 0.12.9 (2021-04-19)
 
 * Fix resolution of imports using local paths not relative to the main document
-* Update to dxCommon 0.2.12
 
 ## 0.12.8 (2021-04-19)